# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Money Planner API",
    "description": "Контракт для iOS и будущего Android-клиента. Даты — YYYY-MM-DD, суммы — decimal-строки с 2 знаками (RUB). Все запросы, кроме /auth/*, — с Authorization: Bearer <accessToken>. Ошибки — единый формат { error: { code, message, status, retryable, requestId, details } }.",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/auth/anonymous": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create Anonymous",
        "operationId": "create_anonymous_v1_auth_anonymous_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnonymousRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh",
        "operationId": "refresh_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/categories": {
      "get": {
        "tags": [
          "categories"
        ],
        "summary": "List Categories",
        "operationId": "list_categories_v1_categories_get",
        "parameters": [
          {
            "name": "includeArchived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Includearchived"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "categories"
        ],
        "summary": "Create Category",
        "operationId": "create_category_v1_categories_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/categories/{category_id}": {
      "patch": {
        "tags": [
          "categories"
        ],
        "summary": "Update Category",
        "operationId": "update_category_v1_categories__category_id__patch",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Category Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "categories"
        ],
        "summary": "Archive Category",
        "operationId": "archive_category_v1_categories__category_id__delete",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/calendar/days": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "Get Days",
        "operationId": "get_days_v1_calendar_days_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarDays"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/calendar/acknowledgements/{day}": {
      "put": {
        "tags": [
          "calendar"
        ],
        "summary": "Acknowledge",
        "operationId": "acknowledge_v1_calendar_acknowledgements__day__put",
        "parameters": [
          {
            "name": "day",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Day"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarDay"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "calendar"
        ],
        "summary": "Unacknowledge",
        "operationId": "unacknowledge_v1_calendar_acknowledgements__day__delete",
        "parameters": [
          {
            "name": "day",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Day"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarDay"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analytics/summary": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Summary",
        "operationId": "get_summary_v1_analytics_summary_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "To"
            }
          },
          {
            "name": "compareFrom",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Comparefrom"
            }
          },
          {
            "name": "compareTo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Compareto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analytics/dynamics": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Get Dynamics",
        "operationId": "get_dynamics_v1_analytics_dynamics_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "To"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Categoryid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DynamicsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnonymousRequest": {
        "properties": {
          "platform": {
            "type": "string",
            "enum": [
              "ios",
              "android"
            ],
            "title": "Platform"
          },
          "appVersion": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Appversion"
          }
        },
        "type": "object",
        "required": [
          "platform",
          "appVersion"
        ],
        "title": "AnonymousRequest"
      },
      "AuthSession": {
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "title": "Userid"
          },
          "isAnonymous": {
            "type": "boolean",
            "title": "Isanonymous"
          },
          "accessToken": {
            "type": "string",
            "title": "Accesstoken"
          },
          "accessTokenExpiresAt": {
            "type": "string",
            "format": "date-time",
            "title": "Accesstokenexpiresat"
          },
          "refreshToken": {
            "type": "string",
            "title": "Refreshtoken"
          },
          "refreshTokenExpiresAt": {
            "type": "string",
            "format": "date-time",
            "title": "Refreshtokenexpiresat"
          }
        },
        "type": "object",
        "required": [
          "userId",
          "isAnonymous",
          "accessToken",
          "accessTokenExpiresAt",
          "refreshToken",
          "refreshTokenExpiresAt"
        ],
        "title": "AuthSession"
      },
      "AveragesOut": {
        "properties": {
          "accountedDays": {
            "type": "integer",
            "title": "Accounteddays"
          },
          "perDay": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^-?\\d+\\.\\d{2}$",
                "example": "1250.00"
              },
              {
                "type": "null"
              }
            ],
            "title": "Perday"
          },
          "weekday": {
            "$ref": "#/components/schemas/BucketOut"
          },
          "weekend": {
            "$ref": "#/components/schemas/BucketOut"
          }
        },
        "type": "object",
        "required": [
          "accountedDays",
          "perDay",
          "weekday",
          "weekend"
        ],
        "title": "AveragesOut"
      },
...
      "CalendarDay": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "expenseCount": {
            "type": "integer",
            "title": "Expensecount"
          },
          "isAcknowledgedEmpty": {
            "type": "boolean",
            "title": "Isacknowledgedempty"
          },
          "isAccounted": {
            "type": "boolean",
            "title": "Isaccounted"
          }
        },
        "type": "object",
        "required": [
          "date",
          "expenseCount",
          "isAcknowledgedEmpty",
          "isAccounted"
        ],
        "title": "CalendarDay"
      },
      "CalendarDays": {
        "properties": {
          "days": {
            "items": {
              "$ref": "#/components/schemas/CalendarDay"
            },
            "type": "array",
            "title": "Days"
          }
        },
        "type": "object",
        "required": [
          "days"
        ],
        "title": "CalendarDays"
      },
      "CategoryCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "emoji": {
            "type": "string",
            "title": "Emoji"
          }
        },
        "type": "object",
        "required": [
          "name",
          "emoji"
        ],
        "title": "CategoryCreate"
      },
      "CategoryList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CategoryOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "CategoryList"
      },
      "CategoryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "emoji": {
            "type": "string",
            "title": "Emoji"
          },
          "kind": {
            "type": "string",
            "enum": [
              "system",
              "custom"
            ],
            "title": "Kind"
          },
          "colorIndex": {
            "type": "integer",
            "title": "Colorindex"
          },
          "sortOrder": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sortorder"
          },
          "archivedAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archivedat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "emoji",
          "kind",
          "colorIndex",
          "sortOrder",
          "archivedAt",
          "createdAt",
          "updatedAt"
        ],
        "title": "CategoryOut"
      },
      "CategorySummaryOut": {
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "title": "Categoryid"
          },
          "total": {
            "type": "string",
            "pattern": "^-?\\d+\\.\\d{2}$",
            "title": "Total",
            "example": "1250.00"
          },
          "share": {
            "type": "number",
            "title": "Share"
          },
          "expenseCount": {
            "type": "integer",
            "title": "Expensecount"
          },
          "averageExpense": {
            "type": "string",
            "pattern": "^-?\\d+\\.\\d{2}$",
            "title": "Averageexpense",
            "example": "1250.00"
          },
          "delta": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^-?\\d+\\.\\d{2}$",
                "example": "1250.00"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta"
          },
          "deltaPercent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deltapercent"
          }
        },
        "type": "object",
        "required": [
          "categoryId",
          "total",
          "share",
          "expenseCount",
          "averageExpense",
          "delta",
          "deltaPercent"
        ],
        "title": "CategorySummaryOut"
      },
      "CategoryUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "emoji": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emoji"
          }
        },
        "type": "object",
        "title": "CategoryUpdate"
      },
      "DataRangeOut": {
        "properties": {
          "firstDate": {
            "type": "string",
            "format": "date",
            "title": "Firstdate"
          },
          "lastDate": {
            "type": "string",
            "format": "date",
            "title": "Lastdate"
          }
        },
        "type": "object",
        "required": [
          "firstDate",
          "lastDate"
        ],
        "title": "DataRangeOut"
      },
      "DynamicsOut": {
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/PointOut"
            },
            "type": "array",
            "title": "Points"
          },
          "averages": {
            "$ref": "#/components/schemas/AveragesOut"
          }
        },
        "type": "object",
        "required": [
          "points",
          "averages"
        ],
        "title": "DynamicsOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PointOut": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "total": {
            "type": "string",
            "pattern": "^-?\\d+\\.\\d{2}$",
            "title": "Total",
            "example": "1250.00"
          }
        },
        "type": "object",
        "required": [
          "date",
          "total"
        ],
        "title": "PointOut"
      },
      "PreviousOut": {
        "properties": {
          "total": {
            "type": "string",
            "pattern": "^-?\\d+\\.\\d{2}$",
            "title": "Total",
            "example": "1250.00"
          }
        },
        "type": "object",
        "required": [
          "total"
        ],
        "title": "PreviousOut"
      },
      "RefreshRequest": {
        "properties": {
          "refreshToken": {
            "type": "string",
            "maxLength": 256,
            "minLength": 16,
            "title": "Refreshtoken"
          }
        },
        "type": "object",
        "required": [
          "refreshToken"
        ],
        "title": "RefreshRequest"
      },
      "SummaryOut": {
        "properties": {
          "total": {
            "type": "string",
            "pattern": "^-?\\d+\\.\\d{2}$",
            "title": "Total",
            "example": "1250.00"
          },
          "previous": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreviousOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategorySummaryOut"
            },
            "type": "array",
            "title": "Categories"
          },
          "dataRange": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DataRangeOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "total",
          "previous",
          "categories",
          "dataRange"
        ],
        "title": "SummaryOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}